#261 GetTableDataPerformance
