+++ components/elements/TagRow.tsx
@@ -2,13 +2,11 @@ import { Tag } from 'antd';
 import { TagProps } from 'antd';
 
 interface TagRowProps {
-  items: { value: string; tagProps?: TagProps }[];
+  items: { value: string | undefined; tagProps?: TagProps }[];
 }
 
 function TagRow({ items }: TagRowProps) {
-  // remove all undefined values from items
   items = items.filter((item) => item.value);
-  console.log(items);
   return (
     <div>
       {items.map(({ value, tagProps }) => (
