+++ components/services/data/GetTableDataPerformance/shared.ts
@@ -1,0 +1,5 @@
+interface SharedBody {
+  allData: boolean;
+  startDate?: string;
+  endDate?: string;
+}
